+++ bapsflib/_hdf/utils/tests/__init__.py
@@ -8,28 +8,12 @@
 # License: Standard 3-clause BSD; see "LICENSES/LICENSE.txt" for full
 #   license terms and contributor agreement.
 #
+__all__ = ['TestBase', 'with_bf']
+
 import unittest as ut
 
 from bapsflib._hdf.maps import FauxHDFBuilder
-from functools import wraps
-
-from ..file import File
-
-
-def with_bf(func):
-    """
-    Context decorator for managing the opening and closing BaPSF HDF5
-    Files :class:`bapsflib._hdf.utils.file.File`.  Intended for use on
-    test methods.
-    """
-    @wraps(func)
-    def wrapper(self, *args, **kwargs):
-        with File(self.f.filename,
-                  control_path='Raw data + config',
-                  digitizer_path='Raw data + config',
-                  msi_path='MSI') as bf:
-            return func(self, bf, *args, **kwargs)
-    return wrapper
+from bapsflib.utils.decorators import with_bf
 
 
 class TestBase(ut.TestCase):
